+++ Package.swift
@@ -65,7 +65,6 @@ let package = Package(
             .product(name: "NIO", package: "swift-nio"),
             .product(name: "NIOHTTP1", package: "swift-nio"),
             .product(name: "NIOWebSocket", package: "swift-nio"),
-            .product(name: "NIOExtras", package: "swift-nio-extras"),
         ]),
         .target(name: "RSocketTCPTransport", dependencies: [
             "RSocketCore",
