+++ Sources/RSocketWebSocketTransport/WSTransport.swift
@@ -19,7 +19,6 @@ import NIO
 import NIOHTTP1
 import NIOWebSocket
 import RSocketCore
-import NIOExtras
 
 /// generates 16 bytes randomly and encodes them as a base64 string as defined in RFC6455 https://tools.ietf.org/html/rfc6455#section-4.1
 /// - Returns: base64 encoded string
@@ -65,58 +64,6 @@ extension WSTransport: TransportChannelHandler {
             }
         )
         return channel.pipeline.addHTTPClientHandlers(withClientUpgrade: config)
-            
             .flatMap { channel.pipeline.addHandler(httpHandler) }
         }
-}
-
-class WebSocketFrameToByteBuffer: ChannelInboundHandler {
-    typealias InboundIn = WebSocketFrame
-    typealias InboundOut = ByteBuffer
-    typealias OutboundOut = WebSocketFrame
-    func channelRead(context: ChannelHandlerContext, data: NIOAny) {
-        let frame = unwrapInboundIn(data)
-        switch frame.opcode {
-        case .continuation:
-            /// we currently do not support WebSocket fragmentation
-            break
-        case .connectionClose:
-            /// TODO: We probably want to handle it the same as if an RSocket close frame and close the connection gracefully
-            break
-        case .ping:
-            pong(context: context, frame: frame)
-        case .pong:
-            /// we never send ping frames, therefore should not receive pong frames
-            break
-        case .text:
-            /// we only support binary frames
-            break
-        case .binary:
-            context.fireChannelRead(wrapInboundOut(frame.unmaskedData))
-        default:
-            /// We handle all opcodes which are defined by WebSocket.
-            /// This should never be reached but WebSocketOpcode is a struct and not an enum and we never no for sure
-            break
-        }
-    }
-    private func pong(context: ChannelHandlerContext, frame: WebSocketFrame) {
-        let responseFrame = WebSocketFrame(fin: true, opcode: .pong, data: frame.unmaskedData)
-        context.write(self.wrapOutboundOut(responseFrame), promise: nil)
-    }
-}
-
-fileprivate func randomMaskingKey() -> WebSocketMaskingKey {
-    let mask = UInt32.random(in: UInt32.min...UInt32.max)
-    return withUnsafeBytes(of: mask) { WebSocketMaskingKey($0)! }
-}
-
-class WebSocketFrameFromByteBuffer: ChannelOutboundHandler {
-    typealias OutboundIn = ByteBuffer
-    typealias OutboundOut = WebSocketFrame
-    func write(context: ChannelHandlerContext, data: NIOAny, promise: EventLoopPromise<Void>?) {
-        let buffer = unwrapOutboundIn(data)
-        let maskKey = randomMaskingKey()
-        let frame = WebSocketFrame(fin: true, opcode: .binary, maskKey: maskKey, data: buffer)
-        context.write(wrapOutboundOut(frame), promise: promise)
-    }
 }
